packages/cli/templates/oauth-proxy/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
python-jose[cryptography]>=3.3.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
